Fixing two small issues with invalid sessions and cookies
I was encountering two small issues with the plugin.

The first was that the first time I visited the application I would get
an AssertionFailed exception.

The second was that after a session was invalidated the application
would create a duplicate session cookie, causing my session information
to get lost.  This happened in Chrome, not sure if that had something
to do with it.  It should now ensure that the cookie that is returned
has the same path and host as the original cookie.

Also revved the version to the next micro plus snapshot.

Conflicts:
	DatabaseSessionGrailsPlugin.groovy
	src/java/grails/plugin/databasesession/SessionProxyFilter.java